+++ src/Public/Export-CompleterSet.ps1
@@ -157,7 +157,7 @@ function Export-CompleterSet
             foreach ($entry in $entriesByPath.Values)
             {
                 $relativePath = [System.IO.Path]::GetRelativePath($outputDirectory, $entry.Path)
-                $writtenPath = if ([System.IO.Path]::IsPathRooted($relativePath)) { $entry.Path } else { $relativePath }
+                $writtenPath = if ([System.IO.Path]::IsPathRooted($relativePath)) { $entry.Path } else { $relativePath.Replace('\', '/') }
 
                 $lines.Add('        @{')
                 $lines.Add("            Path    = '$($writtenPath.Replace("'", "''"))'")